requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
Django==3.2.3
django-extensions==3.1.3
django-sendsms==0.4
django-sms==0.5.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.7.1
docopt==0.6.2
idna==2.10
pipreqs==0.4.10
PyJWT==2.1.0
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
twilio==6.59.0
urllib3==1.26.5
yarg==0.1.9
